+++ src/qwip/database/database.py
@@ -1,5 +1,6 @@
 import functools
 import re
+from contextlib import contextmanager
 from typing import Self
 
 import attrs
@@ -198,6 +199,14 @@ class Database:
                 event.listens_for(engine, "begin")(sqlite_begin)
 
         return engine
+
+    @contextmanager
+    def begin(self):
+        if self.session.in_transaction():
+            yield self.session.get_transaction()
+        else:
+            with self.session.begin() as s:
+                yield s
 
     def connect(self, test: bool = True, timeout: int = 2):
         """Connect to the database.